refactor test probes
Move code to add probes for testing out to a separate
file and add config option to enable/disable.  We
don't want this enable normally but is very useful
when developing GUI/styles for probe. As such its
good to leave in and make it easy to enable when
needed.